+++ Sources/Basics/Collections/IdentifiableSet.swift
@@ -47,11 +47,20 @@ public struct IdentifiableSet<Element: Identifiable>: Collection {
     }
 
     public subscript(id: Element.ID) -> Element? {
-        self.storage[id]
+        get {
+            self.storage[id]
+        }
+        set {
+            self.storage[id] = newValue
+        }
     }
 
     public func index(after i: Index) -> Index {
         Index(storageIndex: self.storage.elements.index(after: i.storageIndex))
+    }
+
+    public mutating func insert(_ element: Element) {
+        self.storage[element.id] = element
     }
 
     public func union(_ otherSequence: some Sequence<Element>) -> Self {