+++ Tests/BuildTests/ClangTargetBuildDescriptionTests.swift
@@ -25,6 +25,27 @@ final class ClangTargetBuildDescriptionTests: XCTestCase {
         XCTAssertFalse(try targetDescription.basicArguments().contains("-w"))
     }
 
+    func testSwiftCorelibsFoundationIncludeWorkaround() throws {
+        let toolchain = MockToolchain(swiftResourcesPath: AbsolutePath("/fake/path/lib/swift"))
+
+        let macosParameters = mockBuildParameters(toolchain: toolchain, triple: .macOS)
+        let linuxParameters = mockBuildParameters(toolchain: toolchain, triple: .arm64Linux)
+        let androidParameters = mockBuildParameters(toolchain: toolchain, triple: .arm64Android)
+
+        let macDescription = try makeTargetBuildDescription("swift-corelibs-foundation",
+                                                            buildParameters: macosParameters)
+        XCTAssertFalse(try macDescription.basicArguments().contains("\(macosParameters.toolchain.swiftResourcesPath!)"))
+
+        let linuxDescription = try makeTargetBuildDescription("swift-corelibs-foundation",
+                                                              buildParameters: linuxParameters)
+        print(try linuxDescription.basicArguments())
+        XCTAssertTrue(try linuxDescription.basicArguments().contains("\(linuxParameters.toolchain.swiftResourcesPath!)"))
+
+        let androidDescription = try makeTargetBuildDescription("swift-corelibs-foundation",
+                                                                buildParameters: androidParameters)
+        XCTAssertTrue(try androidDescription.basicArguments().contains("\(androidParameters.toolchain.swiftResourcesPath!)"))
+    }
+
     func testWarningSuppressionForRemotePackages() throws {
         let targetDescription = try makeTargetBuildDescription("test-warning-supression", usesSourceControl: true)
         XCTAssertTrue(try targetDescription.basicArguments().contains("-w"))
@@ -44,8 +65,8 @@ final class ClangTargetBuildDescriptionTests: XCTestCase {
         )
     }
 
-    private func makeResolvedTarget() throws -> ResolvedTarget {
-        ResolvedTarget(
+    private func makeResolvedTarget() throws -> ResolvedModule {
+        ResolvedModule(
             packageIdentity: .plain("dummy"),
             underlying: try makeClangTarget(),
             dependencies: [],